refactor : Cars 클래스 내에서 이름 중복, 자동차 1대인 경우 예외 처리를 하도록 분리
